Moved tree series mixin as TreeUtilities.
